requirements.txt
# ─── Core AI ──────────────────────────────────────────────────────────────────
anthropic>=0.40.0

# ─── Environment ──────────────────────────────────────────────────────────────
python-dotenv>=1.0.0

# ─── Social Agent ─────────────────────────────────────────────────────────────
requests>=2.31.0

# ─── Calendar Agent ───────────────────────────────────────────────────────────
google-auth>=2.27.0
google-auth-oauthlib>=1.2.0
google-api-python-client>=2.118.0

# ─── Email Agent ──────────────────────────────────────────────────────────────
sendgrid>=6.11.0

# ─── Utils ────────────────────────────────────────────────────────────────────
pydantic>=2.6.0
